#37 respository residente calculo para CH pendente
